requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
Pillow
